requirements.txt
fastapi
uvicorn
google-genai
requests
docling